backend/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
django-cors-headers==3.10.0
django-data-seeder==0.2.0
django-dotenv==1.4.2
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
kombu==5.1.0
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.2
prompt-toolkit==3.0.21
psycopg2==2.9.1
PyJWT==2.3.0
pytz==2021.3
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
cycler==0.10.0
matplotlib==3.5.0
nose==1.3.7
numpy==1.21.4
pandas==1.3.4
patsy==0.4.1
pydot==1.2.3
pyparsing==2.2.1
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.3
six==1.16.0

